% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov_fast.R
\name{distcov.fast}
\alias{distcov.fast}
\title{Calculate an unbiased estimate of the squared distance covariance}
\usage{
distcov.fast(X, Y)
}
\arguments{
\item{X}{numeric vector containing samples}

\item{Y}{numeric vector containing samples}
}
\value{
double giving the square of the distance covariance
}
\description{
Calculate an unbiased estimate of the squared distance covariance
}
\details{
The calculation of this estimate is based on the algorithm in Huo 2014, i.e., it gives an unbiased estimate of the square of the distance covariance. The calculation uses U-statistics and requires only O(n log n) calculations.
}
