% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov_parallel.R
\name{distcov_parallel}
\alias{distcov_parallel}
\title{Calculate the distance covariance 1}
\usage{
distcov_parallel(X, Y, affine = FALSE, bias_corr = TRUE,
  type.X = "sample", type.Y = "sample", metr.X = "euclidean",
  metr.Y = "euclidean", bandwidth = 1)
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix. In the first case, this input can be either a vector of positive length, a matrix with one column or a data.frame with one column. In this case, type.X must be specified as "sample". In the second case, the input must be a distance matrix corresponding to the sample of interest. In this second case, type.X must be "distance".}

\item{Y}{see X.}

\item{affine}{logical; indicates if the affinely transformed distance covariance should be calculated or not.}

\item{bias_corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated.}

\item{type.X}{either "sample" or "distance"; specifies the type of input for X.}

\item{type.Y}{see type.X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance covariance. TO DO: Provide details for this.}

\item{metr.Y}{see metr.X.}

\item{bandwidth}{currently not implemented.}
}
\value{
numeric giving the distance covariance between samples X and Y.
}
\description{
Calculate the distance covariance 1
}
