% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{distcovmatrix}
\alias{distcovmatrix}
\title{Calculate the distance covariance}
\usage{
distcovmatrix(X, bias.corr = TRUE, type.X = "sample",
  metr.X = "euclidean", dcor = FALSE, test = "none", perms = 499L,
  use = "all")
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix. In the first case, this input can be either a vector of positive length, a matrix with one column or a data.frame with one column. In this case, type.X must be specified as "sample". In the second case, the input must be a distance matrix corresponding to the sample of interest. In this second case, type.X must be "distance".}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated.}

\item{type.X}{either "sample" or "distance"; specifies the type of input for X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance covariance. TO DO: Provide details for this.}

\item{use}{: "all" uses all observations, "complete.obs" only uses observations that are complete for all variables, "pairwise.complete.obs" uses pairwise complete observations for computing distance covariances}
}
\value{
numeric giving the distance covariance between samples X and Y.
}
\description{
Calculate the distance covariance
}
