% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{gausskernel}
\alias{gausskernel}
\title{Calculate the Gaussian kernel matrix for a given vector.}
\usage{
gausskernel(X, bandwidth)
}
\arguments{
\item{X}{a numeric vector.}

\item{sigma}{the bandwidth of the kernel.}
}
\value{
The distance matrix corresponding to X.
}
\description{
Calculate the Gaussian kernel matrix for a given vector.
}
