% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{mroot}
\alias{mroot}
\title{Calculate the square root of a semi-positive definite matrix.}
\usage{
mroot(A)
}
\arguments{
\item{A}{square matrix.}
}
\value{
Square matrix that is a square root of A.
}
\description{
Calculate the square root of a semi-positive definite matrix.
}
