% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov_fast.R
\name{PartialSum2D}
\alias{PartialSum2D}
\title{Calculate partial sums}
\usage{
PartialSum2D(X, Y, Z)
}
\arguments{
\item{X}{numeric vector containing samples}

\item{Y}{numeric vector containing samples}

\item{Z}{numeric vector}
}
\value{
numeric vector
}
\description{
Calculate partial sums
}
\details{
This is a help function used in calculating the distance covariance using the algorithm by Huo 2014.
}
